% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevreulProcess-package.R
\docType{package}
\name{chevreulProcess-package}
\alias{chevreulProcess}
\alias{chevreulProcess-package}
\title{chevreulProcess: Tools for managing SingleCellExperiment objects as projects}
\description{
Tools analyzing SingleCellExperiment objects as projects. for input into the Chevreul app downstream. Includes functions for analysis of single cell RNA sequencing data. Supported by NIH grants R01CA137124 and R01EY026661 to David Cobrinik.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whtns/chevreulProcess}
  \item \url{https://whtns.github.io/chevreulProcess/}
  \item Report bugs at \url{https://github.com/cobriniklab/chevreulProcess/issues}
}

}
\author{
\strong{Maintainer}: Kevin Stachelek \email{kevin.stachelek@gmail.com} (\href{https://orcid.org/0000-0003-2085-695X}{ORCID})

Authors:
\itemize{
  \item Bhavana Bhat \email{bhavanabhat29@gmail.com}
}

}
\keyword{internal}
