% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{get_colData}
\alias{get_colData}
\title{Get cell metadata from a given object}
\usage{
get_colData(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
dataframe containing object metadata
}
\description{
Get cell metadata
}
\examples{

data(small_example_dataset)
get_colData(small_example_dataset)
}
