% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{integrate}
\alias{integrate}
\title{Batch Correct Multiple Single Cell Objects}
\usage{
integrate(sce_list, organism = "human", ...)
}
\arguments{
\item{sce_list}{List of two or more SingleCellExperiment objects}

\item{organism}{human or mouse}

\item{...}{extra args passed to sce_reduce_dimensions}
}
\value{
an integrated SingleCellExperiment object
}
\description{
Batch Correct Multiple Single Cell Objects
}
