% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medium_example_dataset}
\alias{medium_example_dataset}
\title{Medium example SingleCellExperiment}
\format{
A SingleCellExperiment with only 200 genes
}
\source{
chevreuldata::human_gene_transcript_sce()
}
\usage{
medium_example_dataset
}
\description{
subset 200 genes from chevreuldata::human_gene_transcript_sce()
}
\keyword{datasets}
