% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{merge_small_sces}
\alias{merge_small_sces}
\title{Merge Small SingleCellExperiment Objects}
\usage{
merge_small_sces(..., k.filter = 50)
}
\arguments{
\item{...}{two or more singlecell objects}

\item{k.filter}{minimum cell number for integration}
}
\value{
a SingleCellExperiment object
}
\description{
Merge Small SingleCellExperiment Objects
}
