% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{reintegrate_sce}
\alias{reintegrate_sce}
\title{Reintegrate (filtered) SingleCellExperiment objects}
\usage{
reintegrate_sce(object, suffix = "", reduction = "PCA", ...)
}
\arguments{
\item{object}{A SingleCellExperiment objects}

\item{suffix}{to be appended to file saved in output dir}

\item{reduction}{to use default is pca}

\item{...}{extra args passed to sce_integrate}
}
\value{
a SingleCellExperiment object
}
\description{
This function takes a SCE object and performs the below steps
\enumerate{
\item split by batch
\item integrate
\item run integration pipeline and save
}
}
