% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{rename_sce}
\alias{rename_sce}
\title{Give a new project name to a SingleCellExperiment object}
\usage{
rename_sce(object, new_name)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{new_name}{New name to assign}
}
\value{
a renamed SingleCellExperiment object
}
\description{
Give a new project name to a SingleCellExperiment object
}
\examples{


data(small_example_dataset)
rename_sce(small_example_dataset, "new_name")
}
