% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_sce}
\alias{save_sce}
\title{Save object with alabaster}
\usage{
save_sce(object, prefix = "unfiltered", proj_dir = getwd())
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{prefix}{a prefix for saving}

\item{proj_dir}{path to a project directory}
}
\value{
a path to SingleCellExperiment object saved
under the alabaster framework
}
\description{
Save object with alabaster
}
