% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{sce_cluster}
\alias{sce_cluster}
\title{Run Louvain Clustering at Multiple Resolutions}
\usage{
sce_cluster(
  object = object,
  resolution = 0.6,
  custom_clust = NULL,
  reduction = "PCA",
  algorithm = 1,
  ...
)
}
\arguments{
\item{object}{A SingleCellExperiment objects}

\item{resolution}{Clustering resolution}

\item{custom_clust}{custom cluster}

\item{reduction}{Set dimensional reduction object}

\item{algorithm}{1}

\item{...}{extra args passed to single cell packages}
}
\value{
a SingleCellExperiment object with louvain clusters
}
\description{
Run Louvain Clustering at Multiple Resolutions
}
\examples{
data(small_example_dataset)
sce_cluster(small_example_dataset)

}
