% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{sce_clustering_workflow}
\alias{sce_clustering_workflow}
\title{Clustering Workflow}
\usage{
sce_clustering_workflow(
  object,
  excluded_cells,
  resolution = seq(0.2, 1, by = 0.2),
  organism = "human",
  experiment_name = "default_experiment",
  ...
)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{excluded_cells}{named list of cells to exclude}

\item{resolution}{resolution(s) to use for clustering cells}

\item{organism}{Organism}

\item{experiment_name}{name of the experiment}

\item{...}{extra args passed to sce_process}
}
\value{
a clustered SingleCellExperiment object
}
\description{
Cluster and Reduce Dimensions of a object
}
\examples{
data("tiny_sce")
tiny_sce |> 
sce_clustering_workflow(experiment_name = "tiny_sce", organism = "human")


}
