% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{sce_preprocess}
\alias{sce_preprocess}
\title{Preprocess Single Cell Object}
\usage{
sce_preprocess(
  object,
  scale = TRUE,
  normalize = TRUE,
  features = NULL,
  legacy_settings = FALSE,
  ...
)
}
\arguments{
\item{object}{Assay to use}

\item{scale}{Perform linear transformation 'Scaling'}

\item{normalize}{Perform normalization}

\item{features}{Identify highly variable features}

\item{legacy_settings}{Use legacy settings}

\item{...}{extra args passed to scaling functions}
}
\value{
a preprocessed SingleCellExperiment object
}
\description{
Performs standard pre-processing workflow for scRNA-seq data
}
\examples{
data("small_example_dataset")
sce_preprocess(small_example_dataset)

}
