% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{set_colData}
\alias{set_colData}
\title{Set cell metadata}
\usage{
set_colData(object, meta)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{meta}{a dataframe containing object metadata}
}
\value{
a SingleCellExperiment object with new colData
}
\description{
Set cell metadata from a given object
}
\examples{

data(small_example_dataset)
new_meta <- data.frame(row.names = colnames(small_example_dataset))
new_meta$example <- "example"
set_colData(small_example_dataset, new_meta)
}
