% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{transcripts_to_genes}
\alias{transcripts_to_genes}
\title{Ensembl Transcript Ids to Gene Symbols}
\usage{
transcripts_to_genes(transcripts)
}
\arguments{
\item{transcripts}{human transcripts}
}
\value{
a vector of gene symbols
}
\description{
Convert ensembl transcript ids to hgnc gene symbols
}
\examples{

NRL_transcripts_hs <-
c("ENST00000359842", "ENST00000470566", "ENST00000465764")

transcripts_to_genes(transcripts = NRL_transcripts_hs)

}
